Bump version; coverage at 73%.

+++ process_improve/batch/preprocessing.py
@@ -579,7 +579,7 @@ def find_reference_batch(
     scaler = MCUVScaler().fit(basewide)
     mcuv = scaler.fit_transform(basewide)
 
-    n_components = np.floor(settings["n_components"])
+    n_components = int(np.floor(settings["n_components"]))
     pca_first = PCA(n_components=n_components).fit(mcuv)
 
     # Excludes all batches with Hotelling's T2 > 90% limit.
